Update PSoC 6 HAL to v1.2.1

+++ targets/TARGET_Cypress/TARGET_PSOC6/psoc6csp/hal/source/cyhal_wdt.c
@@ -101,8 +101,8 @@ extern "C" {
 #define CYHAL_WDT_MAX_IGNORE_BITS 12
 
 typedef struct {
-    uint32_t min_period_ms; // Minimum period in milliseconds that can be represented with this many ignored bits
-    uint32_t round_threshold_ms; // Timeout threshold in milliseconds from which to round up to the minimum period
+    uint16_t min_period_ms; // Minimum period in milliseconds that can be represented with this many ignored bits
+    uint16_t round_threshold_ms; // Timeout threshold in milliseconds from which to round up to the minimum period
     // uint32_t pre_match_cycles; // The number of clock cycles in the first two full counter cycles (before the match value matters)
 } cyhal_wdt_internal_ignore_bits_data_t;
